Comment out editor plugin
The editor plugin is responsible for rendering a cell with the SQL
magic as SQL code. However, it defaults to Python if it's not an SQL
magic cell which is problematic since not all cells are Python. Some
are Bash or markdown. Note, the markdown is only rendered as Python
when you are trying to edit the cell. It's displayed fine when leaving
the editing mode.